refactor: 카드 등록 섹션 컴포넌트명 변경
